Changing center and placing points to one side
This commit adds a couple of commits to try to make this code legible.

The code can also now change the center of the hexagon and can move all points
to one side of the hexagon, specifically for rhombus 1, which is at the top
part of the hexagon.

This will be required when I generate the points for the scenarios.

Cheers.