Hexagon random generator with Arguments
This commit lets the script create all the points for the hexagons as well
as the random location of wireless access points. It accepts command line
arguments making it easy to modify.

For the time being the display is in the terminal. I am thinking of putting
that in a png as well as showing it on the screen.

I have yet to print out the file with the relevant information for the
simulator script.

Enough for today.