Wifi AP selection, producer, consumer and speed
I have implemented the Wifi AP selection, though it is probably not as nice
as it should be. All Wifi APs use the same SSID on the same channel. This is
not good, but it forces the simulation to constantly change AP, which is not
exactly bad.

Consumer, producer and the speed has been also implemented.