Added one more core node to connect server
Slightly modified the code to have one more first level node that works in mesh
so that we can connect the single server node for the scenario.

Also added some more comments so that it will be slightly easier for people
reading the code to understand what I am doing.