Correctly indented code use of StaMApWifiMac
The sta-wifi-mac.cc was indented to follow my indenting rules. I do not think
this follows ns-3 rules, but at least makes it easier for me to read.

The ndn-mobility-random scenario now points to StaMApWifiMac, meaning that any
change to the extension code will directly affect the simulations.

I have created two vectors in sta-map-wifi-mac.cc to store the information I
have about the APs in the vicinity. Barring big changes, the vectors will be
maintained throughout the simulation.