Addition of mac-tx-middle, changes to StaMAC
I am starting to see a pattern in the Wifi simulation of ns3. The addition of
mac-tx-middle is interesting because it doesn't seem to get installed in the
ns3/ndnSIM package. Since I will need it, I am including it as part of the
design.

I also modified sta-map-wifi-mac to maintain all the SSIDs it encounters. I
will have to modify the channel to find out how far away things are but that
is a whole different ballgame.

The implementation of a Wifi that is closer to real life will be much more
painful than I imagined.