Where to put the SSID calculation
Since RSSI is not standarized, it isn't directly included in Ns3.

After reading through the information and flowing the flow chart found here:

http://www2.engr.arizona.edu/~junseok/ns3_wifi.htm

I have decided to make a function to run within StaMApWifiMac to change the
SSID taking into account the distance the node is from the SSID he has seen.
This is not a perfect combination but it is within the scope of Wifi RSSI.

Cheers.