Wifi settings change
After much fighting with the NS-3 code I have the following updates:

802.11n is not completely supported, so setting the standard sometimes leads to
crashes. The standard that seems to be working is the 802.11g which might be
too old for our purposes.

The WifiManager classes have not been completely tested, making a lot of the
errors based on these classes. MinstrelWifiManager has a serious coding error
that has not yet been fixed in the NS-3 bugzilla.

We are now using the ArfWifiManager, which is one of the first managers that
were implemented.