Adding CS size, elimination of PayloadSize
Realized that the PayloadSize of 1024 is resonable for a Data packet. This
means that to get a 0.15 MBps average data rate, we would need to get a
Interest creation frequency that is 0.15 MBps / 1024. This has been added to
the code.

When looking at the average hop count results, it was obvious that the CS size
was too small. Since increasing the Interest rate by 10 fold, it seemed
obvious that the CS size should also increase. I made it possible to change
this parameter via the command line.